aha/requirements.txt
fastapi
uvicorn[standard]
databases[sqlite] # или databases[postgresql]
sqlalchemy
passlib[bcrypt]
pyjwt
redis
python-multipart # если принимаете файлы